requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
backports.zoneinfo==0.2.1
cffi==1.15.1
coverage==7.2.7
decorator==5.1.1
Django==4.2.3
djangorestframework==3.14.0
enum34==1.1.10
executing==1.2.0
ipython==8.12.2
jedi==0.18.2
matplotlib-inline==0.1.6
mimesis==10.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
picture==0.0.1
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
StrEnum==0.4.15
toml==0.10.2
tqdm==4.65.0
traitlets==5.9.0
typing-extensions==4.7.1
wcwidth==0.2.6
